refactor(inline-styles): rename to no-inline-styles
