+++ docs/rule/no-inline-styles.md
@@ -1,4 +1,4 @@
-## inline-styles
+## no-inline-styles
 
 Inline styles are not the best practice because they are hard to maintain and usually make the overall size of the project bigger. This rule forbids the inline styles.
