+++ lib/rules/index.js
@@ -11,7 +11,7 @@ module.exports = {
   'no-outlet-outside-routes': require('./lint-no-outlet-outside-routes'),
   'html-comments': require('./lint-html-comments'),
   'img-alt-attributes': require('./lint-img-alt-attributes'),
-  'inline-styles': require('./lint-inline-styles'),
+  'no-inline-styles': require('./lint-no-inline-styles'),
   'link-rel-noopener': require('./lint-link-rel-noopener'),
   'triple-curlies': require('./lint-triple-curlies'),
   'self-closing-void-elements': require('./lint-self-closing-void-elements'),
